+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Fixed
 * Fixed an issue where map loading would not include some additional models in some cases.
+* Fixed an issue where xc3_tex would not create the output folder for the `edit-wimdo`, `edit-wilay`, and `edit-wifnt` commands.
 
 ## 0.21.0 - 2025-11-25
 ### Added